# docs/api/management-openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "GEOWorkFlow remote management preview",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/management/theme-workspaces/{workspace}/previews": {
            "post": {
                "operationId": "theme-workspaces.preview",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "x-required-scope": "themes:read",
                "x-idempotent": false,
                "x-receipt": false,
                "responses": {
                    "200": {
                        "description": "Successful API envelope"
                    },
                    "401": {
                        "description": "Authentication required"
                    },
                    "403": {
                        "description": "Permission denied"
                    },
                    "409": {
                        "description": "Baseline or request identity conflict"
                    },
                    "201": {
                        "description": "Resource or receipt created"
                    }
                },
                "parameters": [
                    {
                        "name": "workspace",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}